Update fan & creator portfolio queries

+++ pkg/db/query/creator_portfolio.sql
@@ -5,8 +5,7 @@ RETURNING *;
 -- name: GetPortfolioByCreatorID :one
 SELECT *
 FROM creator_portfolio
-WHERE creator_id = $1
-LIMIT 1;
+WHERE creator_id = $1;
 -- name: UpdateCreatorStockQuantity :exec
 UPDATE creator_portfolio
 SET quantity = $3
@@ -14,4 +13,5 @@ WHERE creator_id = $1
   and stock_id = $2;
 -- name: DeleteStockFromCreatorPortfolio :exec
 DELETE FROM creator_portfolio
-WHERE id = $1;
+WHERE stock_id = $2
+  and creator_id = $1;
